Main.java decoupled.
